+++ Dbarone.Net.Database/Core/Buffers/IStorage.cs
@@ -12,7 +12,7 @@ public interface IStorage
     /// Gets the number of pages currently in the storage.
     /// </summary>
     /// <returns></returns>
-    long StoragePageCount();
+    int StoragePageCount();
 
     /// <summary>
     /// Reads a page from disk.
